requirements.txt
keyboard
pywin32
# pyautogui